Split $api.fp.Mapping to separate source file
Also:
* Rename $api.fp.thunk to $api.fp.Thunk
